+++ lib/experimental/editor-settings.php
@@ -21,6 +21,9 @@ function gutenberg_enable_experiments() {
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-dataviews-media-modal' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalDataViewsMediaModal = true', 'before' );
 	}
+	if ( gutenberg_is_experiment_enabled( 'gutenberg-auto-attach-media' ) ) {
+		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalAutoAttachMedia = true', 'before' );
+	}
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-content-only-inspector-fields' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalContentOnlyInspectorFields = true', 'before' );
 	}